edicion de parking db y agregado rentals

+++ web/db/migrate/20160718221833_add_costo_anho_to_parkings.rb
@@ -1,5 +1,5 @@
 class AddCostoAnhoToParkings < ActiveRecord::Migration[5.0]
   def change
-    add_column :parkings, :costo_anho, :double
+    add_column :parkings, :costo_anho, :decimal
   end
 end
